requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
class-registry==2.1.2
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Faker==25.8.0
filters==1.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
numpy==1.26.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.4.post2
